Add Script to generate or update FvUpdate.xml file with Firmware entries
The current FvUpdate.xml was created as part of a proof of concept to demonstrate EFI capsule functionality.
For real HW, this file is much more detailed and can be difficult / error prone to create.
There are already partition files in the https://github.com/qualcomm-linux/qcom-ptool repo for most supported Qualcomm Linux HW and various storage types (emmc and UFS).

Let's introduce a script which parses one of those `partitions.conf` files as a way of generating an initial `FvUpdate.xml` which can then be customized by the user as needed.  The script takes either:
-  a file path to a local partition file
-  parameters specifying the type of HW and storage type in order to choose the reference partition.conf from the online qcom-ptool repo

NOTE: Once the new FvUpdate.xml file is generated, please update the operation to be performed on each firmware entry.  By default the operation type will be set to IGNORE.

Signed-off-by: Venkat Thogaru <quic_thogaru@quicinc.com>